rename loadGuestListCurrent.js to loadGuestList.js to be consistent with function name; move to src directory
